Fixed a couple of issues

+++ src/priority_queue.h
@@ -4,8 +4,8 @@
 using namespace std;
 
 struct Coordinate {
-    uint8_t x;
-    uint8_t y;
+    int x;
+    int y;
 };
 
 typedef struct {
